Extract FrisbeError extension to a file + Rename test methods
